Initial split to frontend (RoR.exe) and backend (RoR_Main.lib)
Presently the backend does everything including windowing, input capture and rendering. That will be done by frontend in the future.
