+++ source/main/Application.h
@@ -51,6 +51,13 @@
 namespace RoR {
 
 // ------------------------------------------------------------------------------------------------
+// Main message loop (see MainMessageLoop.cpp)
+// ------------------------------------------------------------------------------------------------
+
+int StartApplication(int argc, char* argv[]);
+int RunApplicationMessageLoop();
+
+// ------------------------------------------------------------------------------------------------
 // Global definitions and enumerations
 // ------------------------------------------------------------------------------------------------
